udemy-dash/requirements.txt
python_version >= '3.11'
ipykernel
pandas
plotly
dash
python-dotenv

black # Formatter
isort # Formatter
flake8 # Linting
flake8-comprehensions
mypy # Type annoataion linter

# kaleido # to export static plotly images

# scipy
# Course Original Requirements
# alabaster
# asn1crypto
# astroid
# attrs
# Babel
# backcall
# bleach
# certifi
# cffi
# chardet
# click
# cloudpickle
# colorama
# cryptography
# cycler
# dash
# dash-auth
# dash-core-components
# dash-html-components
# dash-renderer
# dash-table
# decorator
# defusedxml
# docutils
# entrypoints
# Flask
# Flask-Compress
# Flask-SeaSurf
# idna
# imagesize
# ipykernel
# ipython
# ipython-genutils
# isort
# itsdangerous
# jedi
# Jinja2
# jsonschema
# jupyter-client
# jupyter-core
# keyring
# kiwisolver
# lazy-object-proxy
# lxml
# MarkupSafe
# matplotlib
# mccabe
# mistune
# nbconvert
# nbformat
# notebook
# numpy
# numpydoc
# packaging
# pandas
# pandas-datareader
# pandocfilters
# parso
# pickleshare
# plotly
# prometheus-client
# prompt-toolkit
# psutil
# pycodestyle
# pycparser
# pyflakes
# Pygments
# pylint
# pyOpenSSL
# pyparsing
# pyrsistent
# PySocks
# python-dateutil
# pytz
# pyzmq
# QtAwesome
# qtconsole
# QtPy
# requests
# retrying
# rope
# rope-py3k
# scipy
# Send2Trash
# six
# snowballstemmer
# Sphinx
# sphinxcontrib-applehelp
# sphinxcontrib-devhelp
# sphinxcontrib-htmlhelp
# sphinxcontrib-jsmath
# sphinxcontrib-qthelp
# sphinxcontrib-serializinghtml
# spyder
# spyder-kernels
# terminado
# testpath
# tornado
# traitlets
# ua-parser
# urllib3
# wcwidth
# webencodings
# Werkzeug
# win-inet-pton
# wincertstore
# wrapt
#pywin32
#pywinpty

